+++ gastromanager-backend/src/main/java/com/kaiho/gastromanager/domain/inventorymovement/api/InventoryMovementServicePort.java
@@ -6,6 +6,6 @@ import java.util.UUID;
 
 public interface InventoryMovementServicePort {
 
-    void recordInventoryMovement(UUID ingredientUuid, int changeQuantity, String reason, Restaurant uuid);
+    void recordInventoryMovement(UUID ingredientUuid, double changeQuantity, String reason, Restaurant uuid);
 
 }
