+++ gastromanager-backend/src/main/java/com/kaiho/gastromanager/domain/order/exception/InsufficientStockException.java
@@ -1,7 +1,7 @@
 package com.kaiho.gastromanager.domain.order.exception;
 
 public class InsufficientStockException extends RuntimeException {
-    public InsufficientStockException(String ingredientName, int requiredQuantity, int availableQuantity) {
+    public InsufficientStockException(String ingredientName, double requiredQuantity, double availableQuantity) {
         super("Insufficient stock for ingredient: " + ingredientName + ". You are requiring " + requiredQuantity + " but there is only " + availableQuantity + " in stock");
     }
 }
